project/requirements.txt
fastapi==0.70.1
uvicorn==0.16.0
asyncpg==0.25.0
tortoise-orm==0.17.7
aerich==0.6.1